refactor: move hot-path task and capability tests out of hotspot files
