+++ crates/app/src/tools/mod.rs
@@ -55,6 +55,8 @@ pub(crate) mod messaging;
 mod payload;
 mod process_exec;
 mod provider_switch;
+#[cfg(test)]
+mod required_capabilities_tests;
 pub mod runtime_config;
 mod session;
 #[cfg(feature = "memory-sqlite")]
@@ -2483,249 +2485,6 @@ mod tests {
             "feishu.calendar.freebusy"
         );
         assert_eq!(canonical_tool_name("file.read"), "file.read");
-    }
-
-    #[test]
-    fn required_capabilities_follow_effective_tool_request() {
-        let direct_file_read = ToolCoreRequest {
-            tool_name: "file.read".to_owned(),
-            payload: json!({"path": "README.md"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_file_read),
-            BTreeSet::from([Capability::InvokeTool, Capability::FilesystemRead])
-        );
-
-        let direct_file_write = ToolCoreRequest {
-            tool_name: "file.write".to_owned(),
-            payload: json!({"path": "notes.txt", "content": "hello"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_file_write),
-            BTreeSet::from([Capability::InvokeTool, Capability::FilesystemWrite])
-        );
-
-        let direct_file_edit = ToolCoreRequest {
-            tool_name: "file.edit".to_owned(),
-            payload: json!({"path": "notes.txt", "old_string": "a", "new_string": "b"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_file_edit),
-            BTreeSet::from([Capability::InvokeTool, Capability::FilesystemWrite])
-        );
-
-        let direct_memory_search = ToolCoreRequest {
-            tool_name: "memory_search".to_owned(),
-            payload: json!({"query": "deploy freeze"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_memory_search),
-            BTreeSet::from([Capability::InvokeTool, Capability::FilesystemRead])
-        );
-
-        let direct_memory_get = ToolCoreRequest {
-            tool_name: "memory_get".to_owned(),
-            payload: json!({"path": "MEMORY.md"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_memory_get),
-            BTreeSet::from([Capability::InvokeTool, Capability::FilesystemRead])
-        );
-
-        let direct_session_search = ToolCoreRequest {
-            tool_name: "session_search".to_owned(),
-            payload: json!({"query": "deploy freeze"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_session_search),
-            BTreeSet::from([Capability::InvokeTool, Capability::MemoryRead])
-        );
-
-        let direct_sessions_list = ToolCoreRequest {
-            tool_name: "sessions_list".to_owned(),
-            payload: json!({"limit": 5}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_sessions_list),
-            BTreeSet::from([Capability::InvokeTool, Capability::MemoryRead])
-        );
-
-        let direct_session_wait = ToolCoreRequest {
-            tool_name: "session_wait".to_owned(),
-            payload: json!({"session_id": "child-session"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_session_wait),
-            BTreeSet::from([Capability::InvokeTool, Capability::MemoryRead])
-        );
-
-        let direct_session_policy_status = ToolCoreRequest {
-            tool_name: "session_tool_policy_status".to_owned(),
-            payload: json!({"session_id": "root-session"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_session_policy_status),
-            BTreeSet::from([Capability::InvokeTool, Capability::MemoryRead])
-        );
-
-        let direct_web_fetch = ToolCoreRequest {
-            tool_name: "web.fetch".to_owned(),
-            payload: json!({"url": "https://example.com"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_web_fetch),
-            BTreeSet::from([Capability::InvokeTool, Capability::NetworkEgress])
-        );
-
-        let direct_web_search = ToolCoreRequest {
-            tool_name: "web.search".to_owned(),
-            payload: json!({"query": "loongclaw"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_web_search),
-            BTreeSet::from([Capability::InvokeTool, Capability::NetworkEgress])
-        );
-
-        let direct_browser_open = ToolCoreRequest {
-            tool_name: "browser.open".to_owned(),
-            payload: json!({"url": "https://example.com"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_browser_open),
-            BTreeSet::from([Capability::InvokeTool, Capability::NetworkEgress])
-        );
-
-        let direct_browser_extract = ToolCoreRequest {
-            tool_name: "browser.extract".to_owned(),
-            payload: json!({"mode": "page_text"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_browser_extract),
-            BTreeSet::from([Capability::InvokeTool])
-        );
-
-        let direct_browser_click = ToolCoreRequest {
-            tool_name: "browser.click".to_owned(),
-            payload: json!({"id": 1}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_browser_click),
-            BTreeSet::from([Capability::InvokeTool, Capability::NetworkEgress])
-        );
-
-        let direct_bash_exec = ToolCoreRequest {
-            tool_name: "bash.exec".to_owned(),
-            payload: json!({"command": "printf ok"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&direct_bash_exec),
-            BTreeSet::from([
-                Capability::InvokeTool,
-                Capability::FilesystemRead,
-                Capability::FilesystemWrite,
-                Capability::NetworkEgress,
-            ])
-        );
-
-        let invoked_file_read = ToolCoreRequest {
-            tool_name: "tool.invoke".to_owned(),
-            payload: json!({
-                "tool_id": "file.read",
-                "lease": "unused",
-                "arguments": {"path": "README.md"}
-            }),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&invoked_file_read),
-            BTreeSet::from([Capability::InvokeTool, Capability::FilesystemRead])
-        );
-
-        let invoked_memory_search = ToolCoreRequest {
-            tool_name: "tool.invoke".to_owned(),
-            payload: json!({
-                "tool_id": "memory_search",
-                "lease": "unused",
-                "arguments": {"query": "deploy freeze"}
-            }),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&invoked_memory_search),
-            BTreeSet::from([Capability::InvokeTool, Capability::FilesystemRead])
-        );
-
-        let invoked_web_fetch = ToolCoreRequest {
-            tool_name: "tool.invoke".to_owned(),
-            payload: json!({
-                "tool_id": "web.fetch",
-                "lease": "unused",
-                "arguments": {"url": "https://example.com"}
-            }),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&invoked_web_fetch),
-            BTreeSet::from([Capability::InvokeTool, Capability::NetworkEgress])
-        );
-
-        let invoked_bash_exec = ToolCoreRequest {
-            tool_name: "tool.invoke".to_owned(),
-            payload: json!({
-                "tool_id": "bash.exec",
-                "lease": "unused",
-                "arguments": {"command": "printf ok"}
-            }),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&invoked_bash_exec),
-            BTreeSet::from([
-                Capability::InvokeTool,
-                Capability::FilesystemRead,
-                Capability::FilesystemWrite,
-                Capability::NetworkEgress,
-            ])
-        );
-
-        let invoked_claw_plan = ToolCoreRequest {
-            tool_name: "tool.invoke".to_owned(),
-            payload: json!({
-                "tool_id": "claw.migrate",
-                "lease": "unused",
-                "arguments": {"mode": "plan", "input_path": "imports/nanobot"}
-            }),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&invoked_claw_plan),
-            BTreeSet::from([Capability::InvokeTool, Capability::FilesystemRead])
-        );
-
-        let invoked_claw_apply = ToolCoreRequest {
-            tool_name: "tool.invoke".to_owned(),
-            payload: json!({
-                "tool_id": "claw.migrate",
-                "lease": "unused",
-                "arguments": {
-                    "mode": "apply",
-                    "input_path": "imports/nanobot",
-                    "output_path": "loongclaw.toml"
-                }
-            }),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&invoked_claw_apply),
-            BTreeSet::from([
-                Capability::InvokeTool,
-                Capability::FilesystemRead,
-                Capability::FilesystemWrite,
-            ])
-        );
-
-        let malformed_invoke = ToolCoreRequest {
-            tool_name: "tool.invoke".to_owned(),
-            payload: json!({"lease": "unused"}),
-        };
-        assert_eq!(
-            required_capabilities_for_request(&malformed_invoke),
-            BTreeSet::from([Capability::InvokeTool])
-        );
     }
 
     #[cfg(feature = "tool-file")]
